PBM-88 Fixed reader, writer and tests

+++ internal/reader/reader_test.go
@@ -1,6 +1,7 @@
 package reader
 
 import (
+	"flag"
 	"fmt"
 	"io"
 	"math/rand"
@@ -18,6 +19,19 @@ import (
 	pb "github.com/percona/percona-backup-mongodb/proto/messages"
 	"gopkg.in/mgo.v2/bson"
 )
+
+var (
+	keepS3Data     bool
+	keepLocalFiles bool
+)
+
+func TestMain(m *testing.M) {
+	flag.BoolVar(&keepS3Data, "keep-s3-data", false, "Do not delete S3 testing bucket and file")
+	flag.BoolVar(&keepLocalFiles, "keep-local-files", false, "Do not files downloaded from the S3 bucket")
+	flag.Parse()
+
+	os.Exit(m.Run())
+}
 
 func TestFSReader(t *testing.T) {
 	dumpFile := "test.dump"
@@ -143,10 +157,10 @@ func TestReaderFromS3(t *testing.T) {
 		var doc bson.M
 		err := bsonReader.UnmarshalNext(&doc)
 		if err != nil {
-			if err == io.EOF {
-				break
+			if err != io.EOF {
+				t.Errorf("Cannot UnmarshalNext: %s", err)
 			}
-			t.Fatalf("Error while reading doc from the reader: %s", err)
+			break
 		}
 		count++
 	}
@@ -195,17 +209,19 @@ func TestDownloadBigFile(t *testing.T) {
 	}
 	fmt.Printf("Created bucket: %s\n", st.Storages["s3-us-west"].S3.Bucket)
 
-	defer awsutils.DeleteBucket(svc, st.Storages["s3-us-west"].S3.Bucket)
-	defer awsutils.EmptyBucket(svc, st.Storages["s3-us-west"].S3.Bucket)
+	if !keepS3Data {
+		defer awsutils.DeleteBucket(svc, st.Storages["s3-us-west"].S3.Bucket)
+		defer awsutils.EmptyBucket(svc, st.Storages["s3-us-west"].S3.Bucket)
+	}
 
-	bw, err := writer.NewBackupWriter(dumpFile, st.Storages["s3-us-west"], pb.CompressionType_COMPRESSION_TYPE_NO_COMPRESSION, pb.Cypher_CYPHER_NO_CYPHER)
+	bw, err := writer.NewBackupWriter(st.Storages["s3-us-west"], dumpFile, pb.CompressionType_COMPRESSION_TYPE_NO_COMPRESSION, pb.Cypher_CYPHER_NO_CYPHER)
 	if err != nil {
 		t.Fatalf("cannot create a new backup writer: %s", err)
 	}
 
 	fmt.Printf("uploading file %s to bucket %s\n", dumpFile, bucket)
 
-	maxSize := int64(6 * 1024 * 1024)
+	maxSize := int64(6 * 1024 * 1024) // Max BSON size
 	chunkSize := int64(64 * 1024)
 	str := strings.Repeat("a", int(chunkSize))
 	size := int64(0)
@@ -229,11 +245,13 @@ func TestDownloadBigFile(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Cannot create tmp file: %s", err)
 	}
-	defer os.Remove(tmpFile)
+	if !keepLocalFiles {
+		defer os.Remove(tmpFile)
+	}
 
 	io.Copy(fh, rdr)
 	fh.Close()
-	rdr.Close()
+	//rdr.Close()
 	fi, err := os.Stat(tmpFile)
 	if err != nil {
 		t.Errorf("Cannot stat file %s: %s", tmpFile, err)
@@ -241,51 +259,6 @@ func TestDownloadBigFile(t *testing.T) {
 	if fi.Size() != size {
 		t.Errorf("invalid file %s size. Want %d, got %d", tmpFile, size, fi.Size())
 	}
-}
-
-func TestDownloadBigDumpFile(t *testing.T) {
-	dumpFile := "2019-01-22T17:47:58Z_rs1.dump"
-	bucket := fmt.Sprintf("pbm-test-bucket-%d", 80831)
-
-	st := &storage.Storages{
-		Storages: map[string]storage.Storage{
-			"s3-us-west": {
-				Type: "s3",
-				S3: storage.S3{
-					Region: "us-west-2",
-					//EndpointURL: "https://minio",
-					Bucket: bucket,
-					Credentials: storage.Credentials{
-						AccessKeyID:     os.Getenv("AWS_ACCESS_KEY_ID"),
-						SecretAccessKey: os.Getenv("AWS_SECRET_ACCESS_KEY"),
-					},
-				},
-				Filesystem: storage.Filesystem{},
-			},
-			"local-filesystem": {
-				Type: "filesystem",
-				Filesystem: storage.Filesystem{
-					Path: "testdata",
-				},
-			},
-		},
-	}
-
-	rdr, err := MakeReader(dumpFile, st.Storages["s3-us-west"], pb.CompressionType_COMPRESSION_TYPE_NO_COMPRESSION, pb.Cypher_CYPHER_NO_CYPHER)
-	if err != nil {
-		t.Fatalf("Cannot create a local filesystem reader: %s", err)
-	}
-
-	tmpFile := filepath.Join(os.TempDir(), "dump.test")
-	fh, err := os.Create(tmpFile)
-	if err != nil {
-		t.Fatalf("Cannot create tmp file: %s", err)
-	}
-	//defer os.Remove(tmpFile)
-
-	io.Copy(fh, rdr)
-	fh.Close()
-	rdr.Close()
 }
 
 func randomBucket() string {
